# Revert "trie: make fullnode children hash calculation concurrently"