This reverts commit 0f7fbb85d6e939510a3e3bb6493a9a332ddfd8e8.

Although the above mentioned commit seem to have resulted in a 50% speed increase for large synthetic tries, running on production code seem to have a large negative impact (up to 30% slowdown for importing frontier). As of yet I do now know from where the slowdown comes from, but it can be immediately seen when importing blocks with `--fakepow`:

Current master:

```
$ geth --datadir=/tmp/bench --cache=512 --fakepow import homestead.rlp 
[...]
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=985.545ms mgasps=0.000 number=2500 hash=dcdb0a…b6c720
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=1.012s    mgasps=0.000 number=5000 hash=5f91e5…ce68d6
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=1.049s    mgasps=0.000 number=7500 hash=1d3fe0…6b8ff0
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=1.044s    mgasps=0.000 number=10000 hash=dc2d93…70a524
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=1.044s    mgasps=0.000 number=12500 hash=73d170…15885c
```

Reverted commit:

```
$ geth --datadir=/tmp/bench --cache=512 --fakepow import homestead.rlp 
[...]
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=551.623ms mgasps=0.000 number=2500 hash=dcdb0a…b6c720
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=542.250ms mgasps=0.000 number=5000 hash=5f91e5…ce68d6
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=537.683ms mgasps=0.000 number=7500 hash=1d3fe0…6b8ff0
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=540.558ms mgasps=0.000 number=10000 hash=dc2d93…70a524
Imported new chain segment  blocks=2500 txs=0 mgas=0.000 elapsed=541.183ms mgasps=0.000 number=12500 hash=73d170…15885c
```

My hunch for the slowdown might be lock contention for very small tries (perhaps the storage tries?). Alternative hunches might be cache coherence problems caused by having to move the tiny node data between CPU cores. Either way, we should investigate further, but in the mean time reverting seems to be a required thing.

Note, the issue doesn't seem to stem from boilerplate modifications as disabling threading is enough to gain back the performance.

Ping @fjl @rjl493456442 @Arachnid.